AP_SmartAudio: Impl of set channel request
AP_SmartAudio: Fixed sincro vtx_status channel change.

Unit test for chan change.

AP_SmartAudio:Fix setting dbm power monkey test to prevent fail when internal state is not properly settled.

AP_SmartAudio: cleaning some crap

AP_Math: Removed crc_dvb_s2_update header, and stored into AP_SmartAudio

AP_Vehicle: Remove ifdef to make include of AP_SmartAudio

AP_SmartAudio: Removed smartaudio_protocol.cpp and included functions into AP_SmartAudio.

AP_SmartAudio: Extract test and setup into examples

AP_SmartAudio: wip tests

AP_SmartAudio: update method which sync with AP_VideoTx implemented and integration into main loop.

AP_SmartAudio: Remove on_boot not needed.

AP_SmartAudio: Updates method retries get settings while not armed if needed.
